backend/requirements.txt
# === Core Framework ===
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic-settings>=2.1.0

# === CATIA V5 COM Bridge ===
pycatia>=0.6.6
PySide6>=6.6.0

# === LLM SDKs (wired in Step 5) ===
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.4.0
